@@ -1,5 +1,6 @@
 import 'dart:ui';
 
+import 'package:dartx/dartx.dart';
 import 'package:hive/hive.dart';
 import 'package:meta/meta.dart';
 import 'package:schulcloud/app/app.dart';
@@ -12,13 +13,13 @@ part 'data.g.dart';
 
 @HiveType(typeId: TypeId.user)
 class User implements Entity<User> {
-  User({
+  const User({
     @required this.id,
     @required this.firstName,
     @required this.lastName,
     @required this.email,
     @required this.schoolId,
-    @required this.displayName,
+    String displayName,
     @required this.avatarInitials,
     @required this.avatarBackgroundColor,
     @required this.permissions,
@@ -28,7 +29,7 @@ class User implements Entity<User> {
         assert(lastName != null),
         assert(email != null),
         assert(schoolId != null),
-        assert(displayName != null),
+        displayName = displayName ?? '$firstName $lastName',
         assert(avatarInitials != null),
         assert(avatarBackgroundColor != null),
         assert(permissions != null),
@@ -62,8 +63,7 @@ class User implements Entity<User> {
   @HiveField(2)
   final String lastName;
 
-  String get name => '$firstName $lastName';
-  String get shortName => '${firstName[0]}. $lastName';
+  String get shortName => '${firstName.chars.first}. $lastName';
 
   @HiveField(3)
   final String email;
@@ -127,7 +127,7 @@ class Root implements Entity<Root> {
       // guessed it — no events are being returned at all 😂
       final jsonResponse = await services.api.get(
         'calendar',
-        parameters: {'all': 'true'},
+        queryParameters: {'all': 'true'},
       ).parseJsonList(isServicePaginated: false);
       return jsonResponse.map((data) => Event.fromJson(data)).toList();
     },
