@@ -274,6 +274,7 @@ class RecurrenceRuleAdapter extends TypeAdapter<RecurrenceRule> {
 }
 
 // Type ids.
+// Used before: 46
 class TypeId {
   static const entity = 71;
   static const id = 40;
@@ -292,15 +293,20 @@ class TypeId {
 
   static const event = 64;
 
-  static const contentType = 46;
-  static const content = 57;
   static const course = 58;
   static const lesson = 59;
+  static const content = 57;
+  static const unsupportedComponent = 73;
+  static const textComponent = 72;
+  static const etherpadComponent = 74;
+  static const nexboardComponent = 75;
+  static const resourcesComponent = 76;
+  static const resource = 77;
 
   static const article = 56;
 
   static const file = 53;
-  static const filePath = 72;
+  static const filePath = 78;
 }
 
 Future<void> initializeHive() async {
@@ -322,10 +328,15 @@ Future<void> initializeHive() async {
     // Calendar module:
     ..registerAdapter(EventAdapter())
     // Courses module:
-    ..registerAdapter(ContentTypeAdapter())
-    ..registerAdapter(ContentAdapter())
     ..registerAdapter(CourseAdapter())
     ..registerAdapter(LessonAdapter())
+    ..registerAdapter(ContentAdapter())
+    ..registerAdapter(UnsupportedComponentAdapter())
+    ..registerAdapter(TextComponentAdapter())
+    ..registerAdapter(EtherpadComponentAdapter())
+    ..registerAdapter(NexboardComponentAdapter())
+    ..registerAdapter(ResourcesComponentAdapter())
+    ..registerAdapter(ResourceAdapter())
     // News module:
     ..registerAdapter(ArticleAdapter())
     // Files module:
