@@ -2,14 +2,18 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:ui';
 
-import 'package:flutter/material.dart';
-import 'package:get_it/get_it.dart';
 import 'package:dartx/dartx.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_deep_linking/flutter_deep_linking.dart';
+import 'package:get_it/get_it.dart';
 import 'package:html/parser.dart';
 import 'package:http/http.dart';
 import 'package:schulcloud/generated/l10n.dart';
 import 'package:url_launcher/url_launcher.dart';
 
+import 'app_config.dart';
+import 'logger.dart';
+import 'services/api_network.dart';
 import 'services/network.dart';
 
 final services = GetIt.instance;
@@ -97,8 +101,25 @@ extension LegenWaitForItDaryString on String {
 
 /// Tries launching a url.
 Future<bool> tryLaunchingUrl(String url) async {
-  if (await canLaunch(url)) {
-    await launch(url);
+  logger.i("Trying to launch url '$url'…");
+  final resolved = Uri.parse(services.config.baseWebUrl).resolve(url);
+  if (resolved.host == services.config.host) {
+    final result = Matcher.path('content/redirect/{id}')
+        .evaluate(PartialUri.fromUri(resolved));
+    if (result.isMatch) {
+      final response = await services.api.head(
+        'content/redirect/${result.parameters['id']}',
+        followRedirects: false,
+      );
+      final redirect = response.headers['location'];
+      logger.d("Resolved content redirect: '$redirect'");
+      return tryLaunchingUrl(redirect);
+    }
+  }
+
+  final string = resolved.toString();
+  if (await canLaunch(string)) {
+    await launch(string);
     return true;
   }
   return false;
